refactor(rank-fusion): extract AbstractDocumentSearcher and open the request seam
DefaultSearcher was both the keyword searcher and the home of the machinery every
searcher in the package needs: the request round-trip, the response-to-document
conversion, highlighting and provenance. SemanticChunkSearcher inherited from it,
which made the two impossible to change independently.

Extract that machinery into AbstractDocumentSearcher and make the concrete searchers
siblings of it. DefaultSearcher keeps its name and behavior, SemanticChunkSearcher
now extends the abstract base directly, and LegacySearcher is added as a stable name
for "keyword search that never participates in fusion".

search() is split into createSearchCondition() + execute(). The seam matters: a
subclass that has to compute request state before the condition can be built - a
query embedding, a set of sub-queries - can now build its own condition and call
execute(), instead of passing that state out of band. sendRequest() takes the
condition for the same reason.

LegacySearcher registers only when rank.fusion.searchers names it. Registering
unconditionally would add it to the fusion set of every deployment that has not
configured the property, because an empty allow-list means "every registered
searcher".

No behavior change: 103 tests in the package pass unmodified apart from three test
doubles that override sendRequest and follow its new signature.
